+++ packages/react-learn-card/src/components/RoundedSquare/constants.ts
@@ -4,12 +4,12 @@ import jobhistoryGraphic from '../../assets/images/walletjobhistory.webp';
 import learningHistoryGraphic from '../../assets/images/backpack.png';
 import skillsGraphic from '../../assets/images/walletskills.webp';
 import achievementsGraphic from '../../assets/images/walletTrophy.png';
-import badgeGraphic from '../../assets/images/social-badge-2.svg';
+import badgeGraphic from '../../assets/images/social-badge-2.png';
 import membershipGraphic from '../../assets/images/membership-graphic.png';
 
 import apple from '../../assets/images/apple.png';
 import accommodationHands from '../../assets/images/Accommodation-Hands.png';
-import experienceMountain from '../../assets/images/experience-mountain.svg';
+import experienceMountain from '../../assets/images/experience-mountain.png';
 
 import MiniTrophyIcon from '../../assets/images/minitrophycolored.svg';
 import MiniGradIcon from '../../assets/images/minigradcapcolored.svg';