+++ packages/plugins/didkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # learn-card-core
 
+## 1.5.1
+
+### Patch Changes
+
+-   Updated dependencies [[`6f0c776840addd052a9df844fefdcb3186c7678d`](https://github.com/learningeconomy/LearnCard/commit/6f0c776840addd052a9df844fefdcb3186c7678d)]:
+    -   @learncard/types@5.6.5
+    -   @learncard/core@9.3.16
+
 ## 1.5.0
 
 ### Minor Changes