+++ packages/plugins/dynamic-loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # learn-card-core
 
+## 1.0.23
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.16
+
 ## 1.0.22
 
 ### Patch Changes